+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Initial release
 - Package: `com.answufeng.ui`
+- Components: StateLayout, TitleBar, SimpleAdapter, MultiTypeAdapter, LoadMoreAdapter, AwDialog, LoadingDialog, AwAnim, AwItemAnimator, RoundLayout, FlowLayout, BadgeView, DividerDecoration